refactor: `SettingLoader.py` モジュールを統合
